requirements.txt
deep_translator==1.11.4
elasticsearch==8.11.0
Flask==3.0.2
Flask_Cors==4.0.0
langchain==0.1.6
langchain_openai==0.0.5
langdetect==1.0.9
mistralai==0.0.12
nltk==3.8.1
pymongo==4.3.3
PyPDF2==3.0.1
python-dotenv==1.0.1
Requests==2.31.0
Werkzeug==3.0.1
gunicorn==21.2.0